+++ app/assets/stylesheets/provider/_commons.scss
@@ -248,10 +248,3 @@ hr, {
     padding: line-height-times(1);
   }
 }
-
-#quick-starts-entry {
-  height: 100%;
-}
-#quick-starts-resources-inner {
-  margin: -24px;
-}
