+++ app/controllers/provider/admin/quick_starts_controller.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class Provider::Admin::QuickStartsController < FrontendController
+  activate_menu :quick_starts
+
+  layout 'provider'
+
+  def show; end
+end
