--- app/controllers/provider/admin/resources_controller.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-class Provider::Admin::ResourcesController < FrontendController
-  activate_menu :resources
-
-  layout 'provider'
-
-  def index; end
-
-  def show; end
-end
