--- app/javascript/quickstarts/index.js
@@ -1,7 +1,0 @@
-import template from './template.yaml'
-
-const resources = [
-    template
-]
-
-export default resources
