--- app/javascript/quickstarts/template.yaml
@@ -1,99 +1,0 @@
-# Additional info: https://docs.openshift.com/container-platform/4.9/web_console/creating-quick-start-tutorials.html
-metadata:
-  name: getting-started-with-quick-starts
-  # you can add additional metadata here
-  instructional: true
-spec:
-  displayName: Getting started with quick starts
-  durationMinutes: 10
-  # Optional type section, will display as a tile on the card
-  type: 
-    text: Type
-    # 'blue' | 'cyan' | 'green' | 'orange' | 'purple' | 'red' | 'grey'
-    color: grey
-  # - The icon defined as a base64 value. Example flow:
-  # 1. Find an .svg you want to use, like from here: https://www.patternfly.org/v4/guidelines/icons/#all-icons
-  # 2. Upload the file here and encode it (output format - plain text): https://base64.guru/converter/encode/image
-  # 3. compose - `icon: data:image/svg+xml;base64,<base64 string from step 2>`
-  # - If empty string (icon: ''), will use a default rocket icon
-  # - If set to null (icon: ~) will not show an icon
-  icon: data:image/svg+xml;base64,PCEtLSBHZW5lcmF0ZWQgYnkgSWNvTW9vbi5pbyAtLT4KPHN2ZyB2ZXJzaW9uPSIxLjEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgd2lkdGg9IjUxMiIgaGVpZ2h0PSI1MTIiIHZpZXdCb3g9IjAgMCA1MTIgNTEyIj4KPHRpdGxlPjwvdGl0bGU+CjxnIGlkPSJpY29tb29uLWlnbm9yZSI+CjwvZz4KPHBhdGggZD0iTTQ0OCA2NHY0MTZoLTMzNmMtMjYuNTEzIDAtNDgtMjEuNDktNDgtNDhzMjEuNDg3LTQ4IDQ4LTQ4aDMwNHYtMzg0aC0zMjBjLTM1LjE5OSAwLTY0IDI4LjgtNjQgNjR2Mzg0YzAgMzUuMiAyOC44MDEgNjQgNjQgNjRoMzg0di00NDhoLTMyeiI+PC9wYXRoPgo8cGF0aCBkPSJNMTEyLjAyOCA0MTZ2MGMtMC4wMDkgMC4wMDEtMC4wMTkgMC0wLjAyOCAwLTguODM2IDAtMTYgNy4xNjMtMTYgMTZzNy4xNjQgMTYgMTYgMTZjMC4wMDkgMCAwLjAxOS0wLjAwMSAwLjAyOC0wLjAwMXYwLjAwMWgzMDMuOTQ1di0zMmgtMzAzLjk0NXoiPjwvcGF0aD4KPC9zdmc+Cg==
-  prerequisites:
-    - You can optionally list prerequisites
-    - Another prerequisite
-    - These prerequisites are also displayed on the introduction step
-  description: |-
-    This description appears on the card in the quick starts catalog.
-  introduction: |-
-    **This introduction is shown at the beginning of the quick start**
-    - It introduces the quick start and lists the tasks within it.
-    - You can view the [source for this quick start](https://github.com/patternfly/patternfly-quickstarts/blob/main/packages/dev/src/quickstarts-data/yaml/template.yaml) for additional help and information
-  tasks:
-    - title: Task 1 title
-      description: |-
-        ## Text
-
-        1. The main body of the task. You can use markdown syntax here to create list items and more.
-
-          This is a paragraph.  
-          This is another paragraph. Add an empty line between paragraphs for line breaks or two spaces at the end.
-        1. For more information on markdown syntax you can visit [this resource](https://www.markdownguide.org/basic-syntax/).
-        1. A <small>limited set</small> of <strong>HTML tags</strong> [are also supported](https://docs.openshift.com/container-platform/4.9/web_console/creating-quick-start-tutorials.html#supported-tags-for-quick-starts_creating-quick-start-tutorials)
-        
-        ## Highlighting
-
-        To enable highlighting, the markdown syntax should contain:
-        - Bracketed link text
-        - The highlight keyword, followed by the ID of the element that you want to animate
-        - The element to be highlighted, needs a `data-quickstart-id` attribute
-
-        **Example**
-        <pre>
-        [Quick starts nav item]{{highlight context-selector}}
-        </pre>
-        will highlight an element with the `data-quickstart-id="context-selector"` attribute
-        
-        ### Code snippets
-
-        The syntax for an inline code snippet contains:
-        - Text between back quotes, followed by `{{copy}}`
-
-        The following text demonstates an inline-copy element `https://github.com/sclorg/ruby-ex.git`{{copy}}
-
-        The syntax for multi-line code snippets:
-        - Text between triple back quotes, followed by `{{copy}}`
-
-        ```
-        oc new-app ruby~https://github.com/sclorg/ruby-ex.git
-        echo "Expose route using oc expose svc/ruby-ex"
-        oc expose svc/ruby-ex
-        ```{{copy}}  
-
-        - Clicking the _Next_ button will display the **Check your work** module.
-
-      # optional - the task's Check your work module
-      review:
-        instructions: |-
-          - Did you complete the task successfully?
-        failedTaskHelp: This task isn’t verified yet. Try the task again.
-      # optional - the task's success and failure messages
-      summary:
-        success: Shows a success message in the task header
-        failed: Shows a failed message in the task header
-    - title: Task 2 title
-      description: |-
-        For more information, the folks at OpenShift wrote a nice guide here:
-        [Creating quick starts](https://docs.openshift.com/container-platform/4.9/web_console/creating-quick-start-tutorials.html)
-
-        Also, don't forget to [check out the source](https://github.com/patternfly/patternfly-quickstarts/blob/main/packages/dev/src/quickstarts-data/yaml/template.yaml) for this quick start for more information!
-
-        Happy writing!
-      review:
-        instructions: |-
-          - Thanks for checking out this quick start!
-        failedTaskHelp: This task isn’t verified yet. Try the task again.
-  conclusion: |-
-    The conclusion appears in the last section of the quick start.
-  # you can link to the next quick start(s) here
-  nextQuickStart:
-    - mas-alert-note-prereq
