+++ app/views/provider/admin/quick_starts/show.html.slim
@@ -1,0 +1,12 @@
+#quick-start-catalog-container
+
+javascript:
+  // Hack of the year, part 2: For QuickStartCatalogPage to work we need to render it as a children
+  // of QuickStartContainer, we can't just render it in another pack here. This moves the node from
+  // within QuickStartContainer to #quick-start-catalog-container
+  document.addEventListener('DOMContentLoaded', () => {
+    const quickStartCatalogContainer = document.getElementById('quick-start-catalog-container')
+    const quickStartCatalogPageWrapper = document.getElementById('quick-start-catalog-page-wrapper')
+
+    quickStartCatalogContainer.after(quickStartCatalogPageWrapper)
+  })
