--- app/views/provider/admin/resources/show.html.slim
@@ -1,1 +1,0 @@
-div#quick-starts-resources
