+++ app/views/shared/provider/_quick_starts_container.html.slim
@@ -1,0 +1,11 @@
+#quick-start-container data-render-catalog=(active_menu == :quick_starts).to_s
+  = javascript_pack_tag 'quick_starts/container'
+
+javascript:
+  document.addEventListener('DOMContentLoaded', () => {
+    // HACK of the year: We need QuickStartContainer to wrap the whole #wrapper for the Drawer to work properly.
+    const wrapperContainer = document.getElementById('wrapper')
+    const quickStartsContainer = document.querySelector('.pfext-quick-start-drawer__body')
+
+    quickStartsContainer.after(wrapperContainer)
+  })
